pyproject.toml
[project]
name = "medical-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "mailjet-rest (>=1.4.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "geopy (>=2.4.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "medical_app", from = "src"}]


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
